chore: extract progress snackbar component from zaken-werkvoorraad component
